[fix] fix a bug
[修复] 修复一个错误

插件Core/__init__.py中的更改修复了一个bug，当无法在插件中找到指令文档时会发出警告并记录错误信息。这个修复将确保正确处理这种情况，避免潜在的问题发生。

插件Core/plugins/_returning_gift.py中的更改调整了导入语句以解决导入错误。这个修复确保正确导入send_email函数，以便在创建返回的礼物时发送电子邮件。

插件Core/plugins/email.py中的更改移动了submit_email函数的导入语句以解决导入错误。这个修复确保正确导入submit_email函数，以便在发送电子邮件时使用。

插件Core/plugins/etm/auto_sign_coupon.py中的更改调整了导入语句以解决导入错误。这个修复确保正确导入send_email函数，以便在自动签到后发送电子邮件。

插件Core/plugins/quickmath_points.py中的更改调整了导入语句以解决导入错误。这个修复确保正确导入send_email函数，以便在快速计算积分插件中发送电子邮件。

插件Core/plugins/redpacket.py中的更改调整了导入语句以解决导入错误。这个修复确保正确导入send_email函数，以便在红包插件中发送电子邮件。

插件Core/plugins/su_reset.py中的更改调整了导入语句以解决导入错误。这个修复确保正确导入send_email函数，以便在su_reset插件中发送电子邮件。
